接入skill_hub代码移入adp
