+++ src/agentscope/app/hub/_skill/__init__.py
@@ -4,7 +4,6 @@
 from ._base import SkillArchive, SkillHubBase
 from ._card import SkillCard, SkillHubPage
 from ._claw_hub import ClawSkillHub
-from ._external_hub import ExternalSkillHub
 
 __all__ = [
     "SkillArchive",
@@ -12,5 +11,4 @@ __all__ = [
     "SkillCard",
     "SkillHubPage",
     "ClawSkillHub",
-    "ExternalSkillHub",
 ]
